Fix configuration errors & typo
Plug install checks in file 'plugs.vim' is not correct since it only
check if config file of the corresponding plug exists. The true way
is to find if the plug is downloaded. So I changes the conditions.

TLDR auto updating is stupid design. I use TLDR for quick inspiration
but it uses auto updating to block user's mind. Well, TLDR offers switch
to disable it, good.

Fixed typo in comment of tmux-256 color in zshrc. Maybe I should remove
it directly? All of my Macs use macOS 14 or higher.